+++ src/shared/components/palette/case-file-view/case-file-view-field.component.ts
@@ -1,0 +1,45 @@
+import { AfterViewInit, Component, ElementRef } from '@angular/core';
+import { fromEvent } from 'rxjs';
+import { map, switchMap, takeUntil } from 'rxjs/operators';
+
+@Component({
+  selector: 'ccd-case-file-view-field',
+  templateUrl: './case-file-view-field.component.html',
+  styleUrls: ['./case-file-view-field.component.scss']
+})
+export class CaseFileViewFieldComponent implements AfterViewInit {
+
+  constructor(private readonly elementRef: ElementRef) {
+  }
+
+  public ngAfterViewInit(): void {
+    const slider = this.elementRef.nativeElement.querySelector('.slider');
+    const documentTreeContainer = this.elementRef.nativeElement.querySelector('.document-tree-container');
+
+    const mousedown$ = fromEvent<MouseEvent>(slider, 'mousedown');
+    const mousemove$ = fromEvent<MouseEvent>(document, 'mousemove');
+    const mouseup$ = fromEvent<MouseEvent>(document, 'mouseup');
+
+    const drag$ = mousedown$.pipe(
+      switchMap(
+        (start) => {
+          const x = start.clientX;
+          const documentTreeContainerWidth = documentTreeContainer.getBoundingClientRect().width;
+          return mousemove$.pipe(map(move => {
+            move.preventDefault();
+            return {
+              dx: move.clientX - x,
+              documentTreeContainerWidth: documentTreeContainerWidth
+            }
+          }),
+          takeUntil(mouseup$));
+        }
+      )
+    );
+
+    drag$.subscribe(pos => {
+      const calculatedWidth = ((pos.documentTreeContainerWidth + pos.dx) * 100) / slider.parentElement.getBoundingClientRect().width;
+      documentTreeContainer.setAttribute('style', `width: ${calculatedWidth}%`);
+    });
+  }
+}