+++ src/shared/components/palette/case-file-view/case-file-view.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CaseFileViewFieldComponent } from './case-file-view-field.component';
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [
+    CaseFileViewFieldComponent
+  ],
+  entryComponents: [
+    CaseFileViewFieldComponent
+  ],
+  exports: [
+    CaseFileViewFieldComponent
+  ]
+})
+export class CaseFileViewModule {}