+++ src/shared/components/palette/case-file-view/index.ts
@@ -1,0 +1,2 @@
+export * from './case-file-view-field.component';
+export * from './case-file-view.module';